Set auto switch between light and dark themes
This utilizes the dark-notify plugin to switch between light and dark
themes based on the system preference. With this I have configured so
Catppuccin Latte is the light theme and Kanagawa is the dark theme.

Kitty also now has support for auto switching between the light and dark
themes based on OS settings as of 0.38.0. This is why I have added the 3
themes under the kitty directory.