Remove copilot for markdown and git commit
I realized it is best if I do not use this for my git commits as it is
usually returning nonsense that has nothing to do with my application.
Even if it is something relevant it should be in my own words. I am
still able to enable these manually by using `Copilot enable`