Fix neovim structure and support Omarchy themes
This makes it so that my neovim config will work better with the setup
that Omarchy is used to. It also restructures the files a bit more into
use case categories and I will still need to clean up a few for better
file specific and LSP support.

The main change is to colour theme which will check the omarchy theme
and match it if we are not on macOS.