Add support for built in LSP config
This makes it so we no longer use the LspZero plugin to manage our
LSP's. A lot of these plugins have now been modified to use this latest
version of `vim.lsp.config` and I needed to migrate. I am still working
on organizing some stuff for on a per language basis but this is a great
start.