Add conform formatting when LSP is unavailable
This makes it so we can use conform as our formatter for tools that we
install in Mason that are only standalone formatters like prettier and
goimports. With this we can also have formatting in visual mode for
specific sections. Conform will also fallback to the LSP functionality
when no formatter is found.